Update package installation script
Prefer online installations if connectivity to the npm registry is available

On Windows, check for the availability of the Expand-Archive cmdlet before attempting to use it. It was introduced in Powershell 5.0 in February 2016, but is not available by default in Windows Server 2012RC2.

+++ src/unix/setup.sh
@@ -1,38 +1,56 @@
 #!/bin/bash
 
-echo '=== Setup ==='
+echo "=== Setup ==="
 
-# Load the correct versions
 pm2_package=$(node src/echo-dependency.js pm2)
 pm2_logrotate_package=$(node src/echo-dependency.js pm2-logrotate)
+
 cache_folder="./.npm_cache";
 cache_archive="./bundle.tar.gz"
 
 # Print out the versions of this package, node, and npm for this host
 node src/bundle/current.js
 
-if [ -f $cache_archive ]; then
+# Check connectivity to registry.npmjs.org
+node src/npm-online.js
 
-  echo 'Cache detected, installing offline..'
+if [ $? -eq 0 ]; then
 
-  # Remove existing cache directory
-  rm -rf $cache_folder
+  echo "Installing packages.."
 
-  # Extract the archive to create a new cache directory
-  tar -xf $cache_archive
+  npm install --global --loglevel=error --no-audit --no-fund $pm2_package
+  npm install --global --loglevel=error --no-audit --no-fund $pm2_logrotate_package
 
-  # Read the bundle information file and compare it to the current host
-  node src/bundle/compare.js
+else 
 
-  npm --offline --cache $cache_folder --optional cache verify
+  echo "Cannot connect to the npm registry. Checking for offline bundle.."
 
-  npm install --global --offline --cache $cache_folder --shrinkwrap false --loglevel=error --audit=false --no-fund $pm2_package
-  npm install --global --offline --cache $cache_folder --shrinkwrap false --loglevel=error --audit=false --no-fund $pm2_logrotate_package
+  if [ -f $cache_archive ]; then
 
-else
-  echo 'Cache not detected, installing online..'
-  npm install --global --loglevel=error --audit=false --no-fund $pm2_package
-  npm install --global --loglevel=error --audit=false --no-fund $pm2_logrotate_package
+    echo "Cache bundle detected. Decompressing.."
+
+    # Remove existing cache directory
+    rm -rf $cache_folder
+
+    # Extract the archive to create a new cache directory
+    tar -xf $cache_archive
+
+    # Read the bundle information file and compare it to the current host
+    node src/bundle/compare.js
+
+    npm --offline --cache $cache_folder --optional cache verify
+
+    echo "Installing packages offline.."
+    npm install --global --offline --cache $cache_folder --shrinkwrap false --loglevel=error --no-audit --no-fund $pm2_package
+    npm install --global --offline --cache $cache_folder --shrinkwrap false --loglevel=error --no-audit --no-fund $pm2_logrotate_package
+
+  else
+
+    echo "Offline bundle not detected. Attempting to install dependencies anyway.."
+
+    npm install --global --loglevel=error --no-audit --no-fund $pm2_package
+    npm install --global --loglevel=error --no-audit --no-fund $pm2_logrotate_package
+  fi
 fi
 
 # Run pm2 on startup
@@ -43,4 +61,4 @@ cd "$(npm config get prefix)/lib/node_modules/pm2-logrotate/"
 pm2 install . --silent
 pm2 save --force
 
-echo '=== Setup Complete ==='
+echo "=== Setup Complete ==="
